Testing Taylor based regularization time
The new method RegCRTBPAtlas.taylorregtime works for 1-dimensional initial data. It still needs to be tested for 0-dimensional data. In this case I'm not even sure what the old algorithm is doing. It may be wrong. It is only used for the decay map sampling at the moment. It needs to be checked before this sampling is trusted.